NB1: Remove unused shared libraries in libfpjni
libfpjni is bloated with a few shared libraries while not utilizing
all of them (ie. the one that got removed in this commit), causing
ELF checks to fail.

Since we never ship/build these libraries anyway and this one never
died on us, remove it from the needed shared libraries.